+++ .gitignore
@@ -1,4 +1,4 @@
 *.pyc
 .coverage
-**/__pycache__/**
+__pycache__/
 venv/
