--- app/psn.py
@@ -1,9 +1,0 @@
-import os
-from flask import Flask
-
-app = Flask(__name__)
-
-
-@app.route('/')
-def index():
-    return "Hello world"
