+++ psn/__init__.py
@@ -1,0 +1,8 @@
+from flask import Flask
+
+from psn.blueprints.base import base
+
+
+app = Flask(__name__)
+
+app.register_blueprint(base)
