+++ psn/blueprints/base.py
@@ -1,0 +1,9 @@
+from flask import Blueprint
+
+
+base = Blueprint('base', __name__, template_folder='templates')
+
+
+@base.route('/')
+def index():
+    return "Hello world"
