+++ requirements.txt
@@ -5,6 +5,7 @@ flask
 nose
 coverage
 codeclimate-test-reporter
+rednose
 
 # utilities
 pycodestyle
