+++ tests/__init__.py
@@ -1,4 +1,5 @@
-import sys,os
+import sys
+import os
 
 import nose
 
@@ -6,4 +7,7 @@ basedir = os.path.dirname(os.path.abspath(__file__))
 
 sys.path.insert(0, basedir + '/../')
 
-from app import psn
+from psn import app
+
+
+client = app.test_client()
