+++ tests/test_base_blueprint.py
@@ -1,0 +1,7 @@
+from . import client
+
+
+def test_index():
+    res = client.get('/')
+
+    assert b'Hello world' in res.data
