--- tests/test_index.py
@@ -1,8 +1,0 @@
-from context import psn
-
-def test_index():
-    client = psn.app.test_client()
-
-    res = client.get('/')
-
-    assert b'Hello world' in res.data
